--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,3 +15,16 @@ log = "0.4"
 
 [dev-dependencies]
 testing_logger = "0.1"
+criterion = { version = "0.6.0", features = ["html_reports"] }
+rayon = "1.10.0"
+minstant = "0.1.7"
+libc = "0.2"
+coarsetime = "0.1"
+
+[features]
+default = []
+lockfree = []
+
+[[bench]]
+name = "log_hz_benchmarks"
+harness = false