--- a/Cargo.toml
+++ b/Cargo.toml
@@ -25,3 +25,8 @@ coarsetime = ["dep:coarsetime"]
 [[bench]]
 name = "log_hz_benchmarks"
 harness = false
+
+[[bench]]
+name = "log_hz_coarsetime_benchmark"
+harness = false
+required-features = ["coarsetime"]