Used atomic var to store timestamp in nanosecond. Also I have added new feature to use coarsetime which is 5x faster than normal instant. Majority of performance gain comes from using faster instant.

```log
log_hz_mutex_vs_lockfree/lockfree_version_coarsetime
                        time:   [8.1911 ns 8.5743 ns 9.1030 ns]
Found 12 outliers among 100 measurements (12.00%)
  2 (2.00%) high mild
  10 (10.00%) high severe

log_hz_mutex_vs_lockfree/mutex_version
                        time:   [42.129 ns 42.278 ns 42.458 ns]
                        change: [−0.3395% +0.0368% +0.4176%] (p = 0.86 > 0.05)
                        No change in performance detected.
Found 11 outliers among 100 measurements (11.00%)
  6 (6.00%) high mild
  5 (5.00%) high severe
  
log_hz_mutex_vs_lockfree/lockfree_version
                        time:   [31.278 ns 32.591 ns 34.390 ns]
                        change: [+1.4293% +4.0472% +7.6982%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  3 (3.00%) high mild
  7 (7.00%) high severe
  ```